Move silhouette pane into a separate file
